Third commit with retrofit network call IService

+++ app/src/main/java/com/example/weatherandroid/Model/Wind.kt
@@ -2,5 +2,5 @@ package com.example.weatherandroid.Model
 
 data class Wind(
     val deg: Int,
-    val speed: Double
+    val speed: Int
 )
